requirements.txt
pygame
pynetworktables